--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/HappyPathAuto.java
@@ -1,8 +1,0 @@
-package org.firstinspires.ftc.teamcode;
-
-/**
- * Created by thund on 10/19/2017.
- */
-
-public class HappyPathAuto {
-}
