--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/put_Arm_Down.java
@@ -1,8 +1,0 @@
-package org.firstinspires.ftc.teamcode;
-
-/**
- * Created by thund on 11/19/2017.
- */
-
-public class put_Arm_Down {
-}
